Added favicon.svg instead of placing the svg as a data url embeded in index.html
This is due to search engines not showing the icon and requiring a separate file. Should work now. Google indexing still pending, but Lepitaxa shows up fine on Bing and Duckduckgo search results.
Dark mode icon color is integrated in the svg file and was tested successfully on Firefox and Edge, therefore assumed to work on all major browsers.