Attempt fix for casing problem with AUT data points, see descr.
Found a problem at the basionym of Ornithoptera victoriae: The AUT starts with a lower case letter, which is not detected by Lepitest. Attempting a fix here, but has to be tested later if it works correctly. Therefore, leaving a lower case letter for exactly this purpose in the data, for now.