Updated the Moon Moths
A. selene + gnoma + artemis still not well resolved, treating gnoma as a ssp. of selene for now, due to it being much closer to it genetically and morphologically than to artemis.
The A. felicis+neidhoeferi clade could potentially contain many cryptic species or ssp., more research needed to resolve them.
Split-off the pine-feeding clade of Actias as a subgenus, well supported by newer phylogenetic data.